agentes2024/dashboard_react_python_supabase/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
python-multipart==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
supabase==2.3.0
pillow==10.2.0
